core: arm: plat-rzg: Add ECC mode checking for shared memory region
When we use ECC setting, the memory and size
of DDR we can use may be changed. So that we must modify them to
compatible with the system(and un-trusted OS).

Otherwise, these macros are used in our platform source only, so that
move to platform private header to avoid warning of undefined macro
value when other source use these pre-define.

Signed-off-by: Huynh Thanh Hung <hung.huynh.wz@renesas.com>
Signed-off-by: hienhuynh <hien.huynh.px@renesas.com>